pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "laser-trim-analyzer"
version = "3.0.2"
description = "Modern QA analysis platform for potentiometer laser trim data"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@company.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Quality Control",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",

    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.1",  # For .xls (Excel 97-2003) files

    # Visualization
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",

    # Database
    "sqlalchemy>=2.0.0",

    # ML/Scientific
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "joblib>=1.3.0",

    # GUI
    "customtkinter>=5.2.0",

    # Utilities
    "pyyaml>=6.0",

    # Memory monitoring (for batch processing on low-RAM systems)
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
    "ipykernel>=6.25.0",  # For Jupyter support
    "notebook>=7.0.0",
]

ml-advanced = [
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "optuna>=3.3.0",  # Hyperparameter optimization
]

performance = [
    "line-profiler>=4.1.0",  # Line-by-line profiling
    "py-spy>=0.3.14",  # Sampling profiler
    "scalene>=1.5.0",  # CPU+GPU+memory profiler
]

all = [
    "laser-trim-analyzer[dev,ml-advanced,performance]",
]

[project.scripts]
laser-trim-analyzer = "laser_trim_analyzer.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
laser_trim_analyzer = ["py.typed", "*.yaml", "*.json"]

# Tool configurations
[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = ["E501"]  # line too long

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
markers = [
    "asyncio: mark test as an asyncio test",
    "slow: mark test as slow running",
    "integration: mark test as integration test", 
    "performance: mark test as performance test",
    "ui: mark test as UI test"
]
python_files = ["test_*.py", "*_test.py"]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*.py"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
